lift struct to a plugin
