--- packages/walt-compiler/src/semantics/map-struct.js
@@ -1,61 +1,0 @@
-// @flow
-import curry from 'curry';
-import invariant from 'invariant';
-import Syntax from '../Syntax';
-import walkNode from '../utils/walk-node';
-import type { NodeType } from '../flow/types';
-import { OBJECT_SIZE, OBJECT_KEY_TYPES, TYPE_OBJECT } from './metadata';
-
-export const getByteOffsetsAndSize = (
-  objectLiteralNode: NodeType
-): [{ [string]: number }, number, { [string]: string }] => {
-  const offsetsByKey = {};
-  const keyTypeMap = {};
-  let size = 0;
-  walkNode({
-    [Syntax.Pair]: keyTypePair => {
-      const { value: key } = keyTypePair.params[0];
-      const { value: typeString } = keyTypePair.params[1];
-      invariant(
-        offsetsByKey[key] == null,
-        `Duplicate key ${key} not allowed in object type`
-      );
-
-      keyTypeMap[key] = typeString;
-      offsetsByKey[key] = size;
-      switch (typeString) {
-        case 'i64':
-        case 'f64':
-          size += 8;
-          break;
-        case 'i32':
-        case 'f32':
-        default:
-          size += 4;
-      }
-    },
-  })(objectLiteralNode);
-
-  return [offsetsByKey, size, keyTypeMap];
-};
-
-const mapStruct = curry(({ userTypes }, [node], _ignore) => {
-  const [offsetsByKey, totalSize, keyTypeMap] = getByteOffsetsAndSize(
-    node.params[0]
-  );
-
-  const struct = {
-    ...node,
-    meta: {
-      ...node.meta,
-      [TYPE_OBJECT]: offsetsByKey,
-      [OBJECT_SIZE]: totalSize,
-      [OBJECT_KEY_TYPES]: keyTypeMap,
-    },
-  };
-
-  userTypes[struct.value] = struct;
-  return struct;
-});
-
-export default mapStruct;
