+++ packages/walt-compiler/src/semantics/metadata.js
@@ -11,7 +11,7 @@ export const TYPE_ARRAY = 'type/array';
 export const TYPE_USER = 'type/user';
 export const TYPE_OBJECT = 'type/object';
 export const TYPE_INDEX = 'TYPE_INDEX';
-export const OBJECT_SIZE = 'object/size';
+export const OBJECT_SIZE = 'OBJECT_SIZE';
 export const TYPE_CAST = 'type/cast';
 export const OBJECT_KEY_TYPES = 'object/key-types';
 export const CLOSURE_TYPE = 'closure/type';
